How to use Framework
Make a folder on your desktop with all of these files in it. The ONLY file we need to edit is the index.html file. The other files are css files and js files that framework incorporates. Also, if you add pictures of gifs you'll have to add them into the folder you placed on your desktop that contains all of the other files as well.